+++ packages/core/config/irving/resolveConfigFilepath.js
@@ -11,7 +11,10 @@ const resolvePackageConfigs = require('./resolvePackageConfigs');
  */
 const resolveConfigFilepath = (configPath) => {
   // Get config paths.
-  const packageConfigPaths = resolvePackageConfigs(configPath, { base: buildContext });
+  const packageConfigPaths = resolvePackageConfigs(
+    configPath,
+    { base: buildContext }
+  );
   const userConfigPath = path.join(buildContext, configPath);
   const configs = packageConfigPaths.concat(userConfigPath);