+++ packages/wordpress/components/blockStyles/blocks/image.js
@@ -1,0 +1,20 @@
+import { css } from 'styled-components';
+import { bodyText, caption, link } from './utils';
+
+export const imageBlock = css`
+  .wp-block-image {
+    ${bodyText};
+
+    figcaption {
+      ${caption};
+
+      a {
+        ${link};
+      }
+    }
+
+    img {
+      height: auto;
+    }
+  }
+`;