+++ packages/wordpress/components/blockStyles/blocks/utils/index.js
@@ -1,0 +1,13 @@
+import bodyText from './bodyText';
+import caption from './caption';
+import link from './link';
+import list from './list';
+import listItem from './listItem';
+
+export {
+  bodyText,
+  caption,
+  link,
+  list,
+  listItem,
+};