+++ packages/wordpress/components/blockStyles/blocks/utils/link.js
@@ -1,0 +1,25 @@
+import { css } from 'styled-components';
+import { siteTheme } from '@irvingjs/styled/utils';
+
+/* eslint-disable max-len, indent */
+const link = css`
+  color: ${siteTheme('blocks.link.color', 'inherit')};
+  border: ${siteTheme('blocks.link.border', 'inherit')};
+  font-family: ${siteTheme(
+    'blocks.link.fontFamily',
+    'Arial, Helvetica, sans-serif'
+  )};
+  font-style: ${siteTheme('blocks.link.fontStyle', 'normal')};
+  font-weight: ${siteTheme('blocks.link.fontWeight', 400)};
+  text-decoration: ${siteTheme('blocks.link.textDecoration', 'underline')};
+
+  &:hover {
+    color: ${siteTheme('blocks.link.hover.color', 'inherit')};
+    text-decoration: ${siteTheme(
+      'blocks.link.hover.textDecoration',
+      'underline'
+    )};
+  }
+`;
+
+export default link;