+++ packages/wordpress/components/blockStyles/index.js
@@ -1,0 +1,11 @@
+import { createGlobalStyle } from 'styled-components';
+import getBlockMap from './blocks';
+
+const blockMap = getBlockMap();
+
+const BlockStylesheet = createGlobalStyle`
+  ${Object.values(blockMap)
+    .map((blockStyles) => blockStyles)}
+`;
+
+export default BlockStylesheet;