+++ packages/wordpress/config/blockStyles.config.js
@@ -1,0 +1,5 @@
+// Default value for user blockMap if unconfigured.
+module.exports = {
+  mergeBlockStyles: true,
+  blockMap: {},
+};